seperate the app logic and render function
