+++ index.html
@@ -36,7 +36,7 @@
       </div>
   <div class="container" id="container">
   </div>
-  <script src="app.js"></script>
+  <script src="app.js" type="module"></script>
 </body>
 
 </html>
